medgemma-report-app/backend/requirements.txt
Flask
python-dotenv
google-cloud-aiplatform
gunicorn
Flask-CORS
google-auth
